pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "denoiser"
version = "0.1.0"
description = "Windows desktop denoising tool for SEM/STEM FA images"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "onnxruntime>=1.21",
  "Pillow>=10",
  "PySide6>=6.7",
  "rosettasciio>=0.13",
  "tifffile>=2024.8.10",
]

[project.optional-dependencies]
build = [
  "pyinstaller>=6.10",
]
dev = [
  "pyinstaller>=6.10",
  "pytest>=8",
]

[project.scripts]
denoiser = "denoiser.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
